+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Take OSN `0.22.10`, previously was on `0.10.10`.
 - [Issue 2](https://github.com/aza547/wow-recorder/issues/2) - Add a monitor selection config option.
 - [Issue 64](https://github.com/aza547/wow-recorder/issues/64), [Issue 60](https://github.com/aza547/wow-recorder/issues/60) - Overhaul async logic causing problems. 
+- [Issue 54](https://github.com/aza547/wow-recorder/issues/54) - Fix to stop recording when leaving arena games with /afk. 
 
 ### Changed
 - Assert that OBS behaves as expected or crash the app, previously we would just log this and get into god knows what error states.